Main menu resolution change

+++ src/TFT_eSPI/TFT_eSPI.h
@@ -649,7 +649,7 @@ class TFT_eSPI : public Print {
 	//////////////////////
   //added functions:
   /////////////////////
-  void drawBitmap(int8_t x, int8_t y, const byte *bitmap, uint16_t color = TFT_BLACK, uint8_t scale = 1);
+  void drawBitmap(int16_t x, int16_t y, const byte *bitmap, uint16_t color = TFT_BLACK, uint8_t scale = 1);
   //void drawBitmap(int8_t x, int8_t y, const byte *bitmap);
   void drawIcon(const unsigned short* icon, int16_t x, int16_t y, uint16_t width, uint16_t height, uint8_t scale = 1);
   void printCenter(const char* text);
